Fix nav menu bug in Licensing 1.3
The new "Calendar" nav button in Licensing module version 1.3 behaves
strangely on hover when logged in as any non-admin user.  This commit
simply moves id="menu-last" to the new last link ("Calendar") in two
places in the code (for users with add/edit privileges, and for all
others), which fixes the problem. This commit also fixes a minor bug to
keep the menu-end image coordinated with the new last button when it is
"on" (or "pressed in").